worker: don’t unnecessarily change directory when running test
subprocess.Popen accepts a `cwd` argument which can be used to change
working directory for the command.  With it, the run_test function no
longer needs to change working directory of the worker process nor does
it need to worry about restoring the old directory.

In particular, because the function had an early return if the outcome was
POSTPONE, it failed to restore the working directory in that case.  This
change fixes that issue.