db: move common DB initialisation code to DB class
All subclasses of the DB class use pretty much the same __init__ method.
SchedulerDB is the only one containing a minor difference in that it passes
commit=False to the DB initialisation method.  To avoid code duplication,
move the common code to DB class.  This makes it possible to get rid of the
__init__ method from UIDB, MasterDB and WorkerDB classes.