scheduler: refactor `request_a_run` endpoint
Add a helper function `request_a_run_impl` which uses a new Failure
exception to indicate any kind of failure.  This simplifies the early
returns as construction of an response object and jsonification of it
is not repeated over and over.

Furthermore, change a few conditions to make the code more concise and
reduce amount of indentation.

Overall, this does not affect behaviour of the code.

Issue: https://github.com/near/nayduck/issues/6