master: don’t specify number of paraller build jobs
The default for Cargo is to use as many build jobs as there are
logical processor cores.  Since our master machines have eight
cores, specifying ‘-j8’ is redundant.  And if the number ever
changes, not having the number hard-coded will make the number
automatically scale if configuration of our master machines change.