ui: optimise get_runs request, i.e. the / endpoint
When backend serves the top 100 runs, it performs a query to fetch the
hundred runs (obviously).  But then, for each of those runs, it
individually executes a query to fetch all builds for the runs.  That
is, a hundred separate queries.  But that’s not all.  For each build
(and there may be multiple per run, often two) it fetches test
statuses statistics.  That’s again at least a hundred queries but
potentially more.

In the end, serving the ‘/’ endpoint make result in hundreds of
queries sent to the database.

Simplify it by performing three separate queries.  One to fetch the
top 100 runs, another one to fetch test statuses for those runs and
final one to fetch build information for all of those runs.  Some more
processing is done in Python now but overall this ends up being quite
noticeably faster.

Issue: https://github.com/near/nayduck/issues/7