worker: clean up test’s temporary directories
Run tests inside of a temporary temporary directory.  That is, create
a temporary directory and set TMPDIR, TEMP and TMP environment
variables to point at it.  Hopefully, the test will respect those
variables and create all temporary files inside of that created
directory.  With that we can delete the temporary directory after the
test is done and thus clean up all files created by the test.

Fixes: https://github.com/near/nayduck/issues/9