db: introduce DB._insert method and rename execute_sql to _execute_sql
Introduce a DB._insert method as a wrapper around _execute_sql which
executes INSERT statements.  It automatically constructs and matches
list of columns, placeholders and values so that it does not need to
be done manually.

While at it, rename DB.execute_sql to DB._execute_sql to indicate that
it’s a protected method and should not be used directly.