ui: optimise schedule request query
Scheduling tests requires adding a rows to runs, builds and tests table.
So far the code would insert the rows one at a time to each of the
tables.  rows table receives a single row either way so that’s not an
issue, but inserting to tests table can easily be done as a single
query.

Refactor request_a_run_impl such that all inserts to the tests table are
done with a single statement.  This should speed scheduling a little
bit.

While at it also remove the `features` column from `tests` table.  It
is never used and in fact is never even set.  In the database its value
is NULL for all rows.

Issue: https://github.com/near/nayduck/issues/7