scheduler: make sure transaction is closed
If an exception is thrown inside of the SchedulerDB::scheduling_a_run
method, a transaction that is started at the beginning of the method
is never closed.  Make sure the transaction is always closed by either
committing or rolling back at the end of the method.

Issue: https://github.com/near/nayduck/issues/13