worker: store raw log size in the database
Rather than prettifying the file size in worker and storing that in
the database, store raw size in the table.  Formatting of the size to
a more human-friendly format can be done when retrieving the value
while at the same time storing prettified string in the database
looses information (rounding the size) and consumes more space.

Issues: https://github.com/near/nayduck/issues/7