master: add `priority` to builds and don’t query `pytest` column
When fetching a pending build to process, prioritise builds which were
requested by people.  To accomplish that, add a `priority` column to builds
table which acts the same way it does in `tests` table.

Note that technically the `priority` column is redundant since it can be
computed by looking at priorities of all tests which depend on a build but
that would make a query run once every five seconds into a join.  It's
better to avoid the join at the cost of a single byte of storage.

While at it, stop fetching `pytest` when getting new build.  The field is
never used and thus unnecessary to compute.

Issue: https://github.com/near/nayduck/issues/7