ui: remove ‘user’ column from runs
The information provided by the ‘user’ column in ‘runs’ table is not
very useful.  It’s username of the person making the commit which can
be retrieved from the repository and does not help with debugging test
failures.

In fact, the only place which attempted to display that information (the
test history view) failed to retrieve the username and no one seemed to
care.

Remove the column from the table.

Issue: https://github.com/near/nayduck/issues/8