worker: don't copy already copies build artefacts
If the test the worker is running uses the same build artefacts as the
one it run previous time, there’s no need to copy the artefacts again
(since the are already on the disk).  For some tests, copying the
files around is actually a significant part of the test execution
time.

Record what was the last build we copied artefacts from such that the
worker can identify situations as above and avoid fetching the files.

Issue: https://github.com/near/nayduck/issues/15