worker: detect if no tests matched test name
If invalid name is passed to as the test name when running expensive
or lib tests the test runs no tests and finishes with success.  This
is intepreted as an IGNORED test but in reality it should be a failed
test since the test that user requested could not be found.

Detect those situation by checking for ‘running 0 tests’ line at the
beginning of standard output.

This helped identify bug in `get_sequential_test_cmd` which did not
pass the name of the test correctly.  Fix that as well.

Issue: https://github.com/near/nayduck/issues/1