db: replace `_execute_sql` with `_exec` method which takes pos. args.
The `_execute_sql` method takes replacement values as a tuple.  This
results in awkward calls when just one value was passed but even with
multiple values the additional parenthesise served no purpose.

Replace the method with `_exec` which is exactly the same expect that
the values are taken as position arguments rather than a single tuple.