workers: do more environment set up in `setup_environ`
Starting the master and worker includes execution of a shell script
which sets up a handful of environment variables.  Move that
initialisation to the `utils.setup_environ` function which already
handles some of the environment variables.

Furthermore, add code for cleaning up unnecessary variables which
may end up in the environment.  This also means remoning database
and Azure configuration from the environment when the service runs.
(Note that Azure configuration is not cleaned from masters'
environments).

Issue: https://github.com/near/nayduck/issues/16