ui: format size on front end side
Rather than converting log file size to string on the back end and
sending the string over, send raw byte size and let the front end
format it in whatever way it wants.  This reduces complexity on the
back end side.

With that also change how log links are formatted a tiny bit.