master: rather than checking % of free space check absolute number
When checking if there’s enough space on the machine, masters will see
whether at least 20% of the data drive is free.  What that 20% means
depends grately on size of the disk and general configuration of the
host.  Change the check to compare with an explicit size of 50GB so
it works the same regardless of size of the disk.

Issue: https://github.com/near/nayduck/issues/16