worker: stop doing s/-/_/ when getting test binary name
The second argument of an expensive test is a test binary name.  Those
happen to use snake_case and as such the worker replaces any dashes it
finds there into underscores.  This is quite redundant however as most
test specs already use snake_case and it only complicates the code for
no benefit.  Remove this behaviour and take the binary name verbatim.