worker: don’t include build type in paths to expensive test binaries
Simplify the code by copying expensive test binaries to `target/expensive`
directory rather than `target_expensive/{build_type}/deps`.  This way, the
test command no longer depends on the build type and the `build_type`
variable does not need to be passed around through so many places.

The same is not done with `target/{build_type}` because many of the tests
depend on path to executables including the build type.