workers: copy only *.wasm files to runtime/near-test-contracts/res
The runtime/near-test-contracts/res directory is a bit weird in that
some of the files in there come from the repository and some of them
are built.

At the moment worker deletes the entire directory and copies data from
master but that results in .gitignore file being skipped (because it
doesn’t match ‘*’).  This isn’t really an issue but it leaves ‘git
checkout’ printing a ‘D’ status line which is just noise in the logs.

Change the code so that only *.wasm files are deleted and copied over.
This avoids the issue with hidden files and also reduces amount of
data being copied over.