worker: retry failed SCP attempts
Add logic to try SCP commands up to three times.  We’ve been seeing
a lot of failures recently caused by SSH server dropping connections,
for example:

    + scp -o StrictHostKeyChecking=no ...
    kex_exchange_identification: read: Connection reset by peer
    Connection reset by ... port 22
    # command finished with exit code 1 after 15 milliseconds

It’s not entirely clear to my why those happen (and I’ve tried
increasing `MaxStartups` setting on the server) so resort to plain old
retries.

While at it configure SSH to use multiplexing so that if multiple SCP
commands are run in succession the client won’t need to reconnect and
reauthenticate each and every time.  The multiplexed connection
remains live for just two seconds which should be enough for a single
test run and at the same time will go back to ‘clean’ state when the
next test is run.