frontend: fix links to non-current branch history
The /api/test/:test_id/history/:branch end point includes a test_id
of one of the tests included in the history so that it’s possible to
explore full history on given branch by viewing history of given test.

Alas, front end overrode that piece of information breaking links to
history on non-current branch by making them lead to history on current
branch.

Fix that.