backend: use testspec for parsing test specifications
Remove code which parses test lines from scheduler module in favour
of the testspec module and convert BackendDB to take advantage of
the TestSpec class.

One effect of this change is that now features are stored in the
‘builds’ column without the ‘--features’ prefix.  Front end and
builders are already prepared for this change so users should not
notice that change.

This also adds a bit more validation to when scheduler parses the
request lines.  Specifically, the ‘expensive’ tests now must have
exactly three arguments (package, executable and test name) while
previously they could be malformed with only two.

Issue: https://github.com/near/nayduck/issues/26