Store ‘timeout’ in a separate column rather than test name
Currently the test timeout is stored in test name as ‘--timout=###’
argument.  This means that any time timeout is adjusted, it is treated
as a new test by the system.  This makes it harder to track history of
the test.

Change the code such that timeout is no longer included in the name
but rather is stored in a separate column in the database.  This way,
changing the timeout does not change the test name and preserves
continuity of test’s history.

Issue: https://github.com/near/nayduck/issues/26