frontend: render pull request links in the commit messages
Match ‘(#xxxx)’ prefixes present in commit message subjects and render
them as links to corresponding pull request.  In run, build and test
views this is simply a matter of changing the ‘#xxxx’ part at the end
into a link.  In all runs and test history views (where commit is
already a link) the ‘(#xxxx)’ is pulled out of the NavLink and turned
into its own link to the PR.