Deprecate ‘mocknet’ tests and add ‘--skip-build’ test category switch
The ‘mocknet’ tests differ in two ways compared to other tests.

First of all, only one mocknet test can run at a time.  This is
implemented by allowing only one of the workers to pick pending
mocknet tests from the database.  This behaviour used to be necessary
because the tests shared common remote infrastructure and would clash
if run concurrently.

Second of all, mocknet tests don’t require build to be performed.
This means that they can start before a build finishes or even if the
build fails.

Because of the second property, mocknet test category has been abused
for fuzz tests which run their own build.  However, the first property
is undesired in that use case.

To improve situation deprecate mocknet tests and introduce
a --skip-build switch which enables the first behaviour on any pytest
test.