backend: group renamed tests into single entry in nightly stats
Many of the tests got renamed over time.  Each time that happens,
they show as completely separate test on the nightly statistics.
This breaks the history and skews the statistics.

Introduce concept of a normalised identifier which tries to apply
such renames to old test names and performs other normalisation
designed to anticipate any renames in the future.

With that, renamed tests are now grouped together in nightly stats.