workers: remove Azure Blob Storage support
We’ve moved from Azure a while back and support for Azure Blob Storage
was present in the code only because it didn’t cost us anything to
keep it there.  However, we now want to add more features to BlobClient
and having to match them in AzureBlobClient is unnecessary distraction.