worker: retry failed tests
To avoid having to deal with flaky tests, if a test fails (or more
specifically, ends up in FAILED or TIMEOUT state) retry it.  This
looses logs of failed runs but at this point this is an acceptable
drawback.

To put some limits, no test will be tried more than three times.
Furthermore, maximum number of tries is dependent on test’s timeout.
No test will be tried for more than an hour (so test with 30 minute
timeout will be tried at most twice).