worker: add support for pausing fuzzer service
There’s currently work being done on a fuzzer service which shares
machines with workers.  The idea is for it to run when CPU is
otherwise idle, i.e. when worker does not run any tests.  To make
this work worker needs to communicate with fuzzer and pause it
whenever it executes a test.  Add support for that.