fuzzers: activate service account at startup
I wrongly believed that gsutil would use machine’s service account when
authenticating with the cloud.  However, that does not happen.  The gcloud
activate-service-account call is required after all so uncomment it.  Note
that the command needs to be called only once but to keep things simpler
we’re calling it each time fuzzer starts.  This should not have negative
effects.